Resolve "product detal rateinfo"

Closes #34

See merge request ronaker/ronaker-customer-android!36

+++ app/src/main/java/com/ronaker/app/data/OrderRepository.kt
@@ -30,7 +30,7 @@ interface OrderRepository {
         token: String?,
         orderSuid: String,
         comment: String,
-        stars: Int
+        stars: Double
     ): Observable<Result<Boolean>>
 }
