+++ app/src/main/java/com/ronaker/app/data/network/request/ProductRateRequestModel.kt
@@ -3,7 +3,7 @@ package com.ronaker.app.data.network.request
 import com.google.gson.annotations.SerializedName
 
 
-data class ProductRateRequestModel(@SerializedName("stars") val stars: Int,
+data class ProductRateRequestModel(@SerializedName("stars") val stars: Double,
                                    @SerializedName("comment") val comment: String)
